+++ cms-client/pages/products/[productId]/variants/[variantId].tsx
@@ -9,10 +9,12 @@ import Media from '@/components/variants/blocks/Media'
 import VariantList from '@/components/variants/blocks/VariantList'
 import { useDeleteVariantMutation, useGetVariantByIdQuery, useUpdateVariantMutation } from '@/services/variantService'
 import { IErrorResponse, VariantCreateRequest, VariantUpdateRequest } from '@/types/api'
+import useConfirm from '@/hooks/useConfirm'
 
 
 function Variant() {
     const router = useRouter()
+    const { show } = useConfirm()
 
     const { isError, error, isLoading, data } = useGetVariantByIdQuery({ variantId: Number(router.query.variantId) }, { skip: router.query.variantId === undefined })
     const [updateVariant, { isSuccess: isUpdateVariantSuccess, isError: isUpdateVariantError, error: updateVariantError }] = useUpdateVariantMutation()
@@ -59,10 +61,14 @@ function Variant() {
     }
 
     const onVariantDelete = async () => {
-        const result = await deleteVariant({ variantId: Number(router.query.variantId) }).unwrap();
-        if (result.success === true) {
-            setChanges({})
-            router.push(`/products/${router.query.productId}`)
+        const isConfirmed = await show("Подтверждение", "Подтвердите удаление варианта")
+
+        if (isConfirmed === true) {
+            const result = await deleteVariant({ variantId: Number(router.query.variantId) }).unwrap();
+            if (result.success === true) {
+                setChanges({})
+                router.push(`/products/${router.query.productId}`)
+            }
         }
     }
